Move common functions into common.php file
